requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
redis==4.5.4
mysql-connector-python==8.0.28
yt-dlp==2023.3.4
whisperx==3.1.1
transformers==4.28.1
torch==2.0.0
python-multipart==0.0.6
